CommandSet: break out nested classes to separate files
