+++ src/System.Windows.Forms.Design/src/System/Windows/Forms/Design/CommandSet.cs
@@ -21,7 +21,7 @@ namespace System.Windows.Forms.Design;
 ///  that are common to both functions.
 /// </summary>
 /// <internalonly/>
-internal class CommandSet : IDisposable
+internal partial class CommandSet : IDisposable
 {
     protected ISite site;
     private readonly CommandSetItem[] _commandSet;
@@ -3531,376 +3531,6 @@ internal class CommandSet : IDisposable
         if (tabIndexCollision)
         {
             componentControl.TabIndex = nextTabIndex;
-        }
-    }
-
-    /// <internalonly/>
-    /// <summary>
-    ///  We extend MenuCommand for our command set items.  A command set item
-    ///  is a menu command with an added delegate that is used to determine the
-    ///  flags for the menu item.  We have different classes of delegates here.
-    ///  For example, many  menu items may be enabled when there is at least
-    ///  one object selected, while others are only enabled if there is more than
-    ///  one object or if there is a primary selection.
-    /// </summary>
-    protected class CommandSetItem : MenuCommand
-    {
-        private readonly EventHandler _statusHandler;
-        private readonly IEventHandlerService _eventService;
-        private readonly IUIService? _uiService;
-
-        private readonly CommandSet? _commandSet;
-        private static readonly Dictionary<EventHandler, StatusState> s_commandStatusHash = new();       // Dictionary of the command statuses we are tracking.
-        private bool _updatingCommand; // flag we set when we're updating the command so we don't call back on the status handler.
-
-        public CommandSetItem(CommandSet commandSet,
-            EventHandler statusHandler,
-            EventHandler invokeHandler,
-            CommandID id,
-            IUIService? uiService) : this(commandSet,
-            statusHandler,
-            invokeHandler,
-            id,
-            optimizeStatus: false,
-            uiService)
-        {
-        }
-
-        public CommandSetItem(CommandSet commandSet,
-            EventHandler statusHandler,
-            EventHandler invokeHandler,
-            CommandID id) : this(commandSet,
-            statusHandler,
-            invokeHandler,
-            id,
-            optimizeStatus: false,
-            uiService: null)
-        {
-        }
-
-        public CommandSetItem(CommandSet commandSet,
-            EventHandler statusHandler,
-            EventHandler invokeHandler,
-            CommandID id,
-            bool optimizeStatus) : this(commandSet,
-            statusHandler,
-            invokeHandler,
-            id,
-            optimizeStatus,
-            uiService: null)
-        {
-        }
-
-        /// <summary>
-        ///  Creates a new CommandSetItem.
-        /// </summary>
-
-        // Per SBurke...
-        public CommandSetItem(CommandSet commandSet,
-            EventHandler statusHandler,
-            EventHandler invokeHandler,
-            CommandID id,
-            bool optimizeStatus,
-            IUIService? uiService)
-            : base(invokeHandler, id)
-        {
-            _uiService = uiService;
-            _eventService = commandSet._eventService;
-            _statusHandler = statusHandler;
-
-            // when we optimize, it's because status is fully based on selection.
-            // so what we do is only call the status handler once per selection change to prevent
-            // doing the same work over and over again.  we do this by hashing up the command statuses
-            // and then filling in the results we get, so we can easily retrieve them when
-            // the selection hasn't changed.
-            //
-            if (optimizeStatus && statusHandler is not null)
-            {
-                // we use this as our sentinel of when we're doing this.
-                //
-                _commandSet = commandSet;
-
-                //
-                // UNDONE:CommandSetItem is put in a static dictionary, and CommandSetItem
-                // references CommandSet, CommandSet reference FormDesigner. If we don't
-                // remove the CommandSetItem from the static dictionary, FormDesigner is
-                // leaked. This demonstrates a bad design. We should not keep a static
-                // dictionary for all the items, instead, we should keep a dictionary per
-                // Designer. When designer is disposed, all command items got disposed
-                // automatically. However, at this time, we would pick a simple way with
-                // low risks to fix this.
-                //
-                // if this handler isn't already in there, add it.
-                //
-                if (!s_commandStatusHash.TryGetValue(statusHandler, out StatusState? state))
-                {
-                    state = new StatusState();
-                    s_commandStatusHash.Add(statusHandler, state);
-                }
-
-                state._refCount++;
-            }
-        }
-
-        /// <summary>
-        /// Checks if the status for this command is valid, meaning we don't need to call the status handler.
-        /// </summary>
-        private bool CommandStatusValid
-        {
-            get
-            {
-                // check to see if this is a command we have hashed up and if it's version stamp
-                // is the same as our current selection version.
-                //
-                if (_commandSet is not null && s_commandStatusHash.TryGetValue(_statusHandler, out StatusState? state))
-                {
-                    if (state.SelectionVersion == _commandSet.SelectionVersion)
-                    {
-                        return true;
-                    }
-                }
-
-                return false;
-            }
-        }
-
-        /// <summary>
-        /// Applies the cached status to this item.
-        /// </summary>
-        private void ApplyCachedStatus()
-        {
-            if (_commandSet is not null && s_commandStatusHash.TryGetValue(_statusHandler, out StatusState? state))
-            {
-                try
-                {
-                    // set our our updating flag so it doesn't call the status handler again.
-                    //
-                    _updatingCommand = true;
-
-                    // and push the state into this command.
-                    //
-                    state.ApplyState(this);
-                }
-                finally
-                {
-                    _updatingCommand = false;
-                }
-            }
-        }
-
-        /// <summary>
-        ///  This may be called to invoke the menu item.
-        /// </summary>
-        public override void Invoke()
-        {
-            // We allow outside parties to override the availability of particular menu commands.
-            //
-            try
-            {
-                if (_eventService is not null)
-                {
-                    IMenuStatusHandler? msh = (IMenuStatusHandler?)_eventService.GetHandler(typeof(IMenuStatusHandler));
-                    if (msh is not null && msh.OverrideInvoke(this))
-                    {
-                        return;
-                    }
-                }
-
-                base.Invoke();
-            }
-            catch (Exception e)
-            {
-                _uiService?.ShowError(e, string.Format(SR.CommandSetError, e.Message));
-
-                if (e.IsCriticalException())
-                {
-                    throw;
-                }
-            }
-        }
-
-        ///<summary>
-        /// Only pass this down to the base when we're not doing the cached update.
-        ///</summary>
-        protected override void OnCommandChanged(EventArgs e)
-        {
-            if (!_updatingCommand)
-            {
-                base.OnCommandChanged(e);
-            }
-        }
-
-        ///<summary>
-        /// Saves the status for this command to the statusstate that's stored in the dictionary
-        /// based on our status handler delegate.
-        ///</summary>
-        private void SaveCommandStatus()
-        {
-            if (_commandSet is not null)
-            {
-                // see if we need to create one of these StatusState dudes.
-                //
-                if (!s_commandStatusHash.TryGetValue(_statusHandler, out StatusState? state))
-                {
-                    state = new StatusState();
-                }
-
-                // and save the enabled, visible, checked, and supported state.
-                //
-                state.SaveState(this, _commandSet.SelectionVersion);
-            }
-        }
-
-        /// <summary>
-        ///  Called when the status of this command should be re-queried.
-        /// </summary>
-        public void UpdateStatus()
-        {
-            // We allow outside parties to override the availability of particular menu commands.
-            //
-            if (_eventService is not null)
-            {
-                IMenuStatusHandler? msh = (IMenuStatusHandler?)_eventService.GetHandler(typeof(IMenuStatusHandler));
-                if (msh is not null && msh.OverrideStatus(this))
-                {
-                    return;
-                }
-            }
-
-            if (_statusHandler is not null)
-            {
-                // if we need to update our status,
-                // call the status handler.  otherwise,
-                // get the cached status and push it into this
-                // command.
-                //
-                if (!CommandStatusValid)
-                {
-                    try
-                    {
-                        _statusHandler(this, EventArgs.Empty);
-                        SaveCommandStatus();
-                    }
-                    catch
-                    {
-                    }
-                }
-                else
-                {
-                    ApplyCachedStatus();
-                }
-            }
-        }
-
-        /// <summary>
-        /// Remove this command item from the static dictionary to avoid leaking this object.
-        /// </summary>
-        public virtual void Dispose()
-        {
-            if (s_commandStatusHash.TryGetValue(_statusHandler, out StatusState? state))
-            {
-                state._refCount--;
-                if (state._refCount == 0)
-                {
-                    s_commandStatusHash.Remove(_statusHandler);
-                }
-            }
-        }
-
-        /// <summary>
-        /// This class saves the state for a given command.  It keeps track of the results
-        /// of the last status handler invocation and what "selection version" that happened on.
-        /// </summary>
-        private class StatusState
-        {
-            // these are the command's possible values.
-            //
-            private const int Enabled = 0x01;
-            private const int Visible = 0x02;
-            private const int Checked = 0x04;
-            private const int Supported = 0x08;
-            private const int NeedsUpdate = 0x10;
-            private int _statusFlags = NeedsUpdate; // our flags.
-
-            // Multiple CommandSetItem instances can share a same status handler within a designer host.
-            // We use a simple ref count to make sure the CommandSetItem can be properly removed.
-            internal int _refCount;
-
-            /// <summary>
-            /// Just what it says...
-            /// </summary>
-            public int SelectionVersion { get; private set; }
-
-            /// <summary>
-            /// Pushes the state stored in this object into the given command item.
-            /// </summary>
-            internal void ApplyState(CommandSetItem item)
-            {
-                Debug.Assert((_statusFlags & NeedsUpdate) != NeedsUpdate, "Updating item when StatusState is not valid.");
-
-                item.Enabled = ((_statusFlags & Enabled) == Enabled);
-                item.Visible = ((_statusFlags & Visible) == Visible);
-                item.Checked = ((_statusFlags & Checked) == Checked);
-                item.Supported = ((_statusFlags & Supported) == Supported);
-            }
-
-            /// <summary>
-            /// Updates this status object  with the state from the given item,
-            /// and saves the selection version.
-            /// </summary>
-            internal void SaveState(CommandSetItem item, int version)
-            {
-                SelectionVersion = version;
-                _statusFlags = 0;
-                if (item.Enabled)
-                {
-                    _statusFlags |= Enabled;
-                }
-
-                if (item.Visible)
-                {
-                    _statusFlags |= Visible;
-                }
-
-                if (item.Checked)
-                {
-                    _statusFlags |= Checked;
-                }
-
-                if (item.Supported)
-                {
-                    _statusFlags |= Supported;
-                }
-            }
-        }
-    }
-
-    /// <internalonly/>
-    /// <summary>
-    ///  The immediate command set item is used for commands that cannot be cached.  Commands
-    ///  such as Paste that get outside stimulus cannot be cached by our menu system, so
-    ///  they get an ImmediateCommandSetItem instead of a CommandSetItem.
-    /// </summary>
-    protected class ImmediateCommandSetItem : CommandSetItem
-    {
-        /// <summary>
-        ///  Creates a new ImmediateCommandSetItem.
-        /// </summary>
-        public ImmediateCommandSetItem(CommandSet commandSet, EventHandler statusHandler, EventHandler invokeHandler, CommandID id, IUIService? uiService)
-        : base(commandSet, statusHandler, invokeHandler, id, uiService)
-        {
-        }
-
-        /// <summary>
-        ///  Overrides OleStatus in MenuCommand to invoke our status handler first.
-        /// </summary>
-        public override int OleStatus
-        {
-            get
-            {
-                UpdateStatus();
-                return base.OleStatus;
-            }
         }
     }
